requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.4
python-dotenv>=1.0.1
anthropic>=0.40.0
openai>=1.30.1
httpx>=0.27.0
esprima>=4.0.1